Move GDB to riscv
