+++ src/main/java/li/cil/sedna/riscv/R5CPUTemplate.java
@@ -11,8 +11,8 @@ import li.cil.sedna.api.device.rtc.RealTimeCounter;
 import li.cil.sedna.api.memory.MappedMemoryRange;
 import li.cil.sedna.api.memory.MemoryAccessException;
 import li.cil.sedna.api.memory.MemoryMap;
-import li.cil.sedna.gdbstub.CPUDebugInterface;
-import li.cil.sedna.gdbstub.Watchpoint;
+import li.cil.sedna.riscv.gdbstub.CPUDebugInterface;
+import li.cil.sedna.riscv.gdbstub.Watchpoint;
 import li.cil.sedna.instruction.InstructionDefinition.Field;
 import li.cil.sedna.instruction.InstructionDefinition.Instruction;
 import li.cil.sedna.instruction.InstructionDefinition.InstructionSize;
